#5256: create service weather get data, weather cache context, configure caching of weather block
